+++ controllers/user.js
@@ -1,11 +1,11 @@
 // Dependencies
-const User = require('../models/UserModel');
-const Profile = require('../models/ProfileModel');
-const Subscription = require('../models/SubscriptionModel');
-const Project = require('../models/ProjectModel');
+const User = require('../models/user');
+const Profile = require('../models/profile');
+const Subscription = require('../models/subscription');
+const Project = require('../models/project');
 const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
-const Email = require('../helpers/email_helpers');
+const Email = require('../helpers/email');
 
 // Models
